Cleaning up some code
Cleaned up code, changed script name, changed all publics to private and serialized fields, commeted through code.
